rename fava.typing to fava.context
