+++ fava/application.py
@@ -45,7 +45,7 @@ from fava.core.helpers import FavaAPIException
 from fava.help import HELP_PAGES
 from fava.json_api import json_api
 from fava.serialisation import serialise
-from fava.typing import g
+from fava.context import g
 from fava.util import resource_path
 from fava.util import send_file_inline
 from fava.util import setup_logging
