+++ fava/context.py
@@ -1,4 +1,5 @@
 """Specify types for the flask application context."""
+
 from typing import Any
 from typing import Optional
